Abstracted current gpt responders to their own module (#178)
See #195  for future work. Full commit list

* abstracted current gpt responders to their own module

* updated wo_requirements to install lo_gpt

* added a comment about how we ought to handle responders in the future

* updated readme with future work
