+++ modules/lo_gpt/setup.cfg
@@ -1,0 +1,10 @@
+[metadata]
+name = Learning Observer GPT
+description = Use this to iterface with GPT models.
+
+[options]
+packages = lo_gpt
+
+[options.entry_points]
+lo_modules =
+   lo_gpt = lo_gpt.module
